+++ home_feature/lib/home_feature.dart
@@ -1,2 +1,2 @@
-export 'package:presentation/presentation.dart' show HomePage, AddTaskPage;
+export 'package:presentation/presentation.dart' show HomePage, EditTaskPage;
 export 'package:home_feature/dependency_injection.dart';
