+++ home_feature/presentation/lib/pages/edit_task_page/local_widgets/add_button.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 
-class AddButton extends StatelessWidget {
+class FinishButton extends StatelessWidget {
   final VoidCallback onPressed;
 
-  const AddButton({
+  const FinishButton({
     super.key,
     required this.onPressed,
   });
