+++ home_feature/presentation/lib/presentation.dart
@@ -1,7 +1,7 @@
 export 'package:presentation/pages/home_page/home_page.dart';
-export 'package:presentation/pages/add_task_page/add_task_page.dart';
+export 'package:presentation/pages/edit_task_page/edit_task_page.dart';
 
-export 'package:presentation/presenters/viewmodels/add_viewmodel.dart';
+export 'package:presentation/presenters/viewmodels/edit_viewmodel.dart';
 export 'package:presentation/presenters/viewmodels/home_viewmodel.dart';
 
 export 'package:presentation/constants/app_constants.dart';
