+++ lib/src/core/routes/app_routes.dart
@@ -28,7 +28,7 @@ final class AppRoutes {
       final DateTime initialDate =
           DateFormat('yyyy/MM/dd').parse(state.pathParameters['start_date']!);
 
-      return AddTaskPage(
+      return EditTaskPage(
         initialDate: initialDate,
         pageNavigator: () => context.pop(),
       );
